backend/requirements.txt
fastapi[standard]>=0.128.0,<0.129.0
sqlalchemy>=2.0.46,<2.1.0
aiosqlite>=0.21.0,<0.22.0
alembic>=1.18.0,<1.19.0
pydantic>=2.12.0,<2.13.0
python-dateutil>=2.9.0,<3.0.0
uvicorn[standard]>=0.34.0,<0.35.0
